move clearComplete function to toDoStorage file
