+++ src/index.js
@@ -1,19 +1,15 @@
 import './style.css';
 import { addIcons, toDos, add } from './modules/ui.js';
+import TodoStorage from './modules/toDoStorage.js';
 
-const clearCompleted = () => {
-  let toDoList = JSON.parse(localStorage.getItem('allEntries'));
-  toDoList = toDoList.filter((toDo) => toDo.completed === false);
-
-  for (let i = 0; i < toDoList.length; i += 1) {
-    toDoList[i].index = i + 1;
-  }
-  localStorage.setItem('allEntries', JSON.stringify(toDoList));
-  toDos();
-};
+const clear = () =>{
+  const storage = new TodoStorage();
+  storage.clearCompleted()
+  toDos()
+}
 
 addIcons();
 toDos();
 document.querySelector('#addBtn').addEventListener('click', add);
 document.querySelector('#refreshBtn').addEventListener('click', toDos);
-document.querySelector('#clearBtn').addEventListener('click', clearCompleted);
+document.querySelector('#clearBtn').addEventListener('click', clear);
